backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
pymysql==1.1.0
mysqlclient==2.2.0
django-cors-headers==4.3.0
python-dotenv==1.0.0